backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.1,<4.0
gunicorn>=22.0,<23.0
celery>=5.4,<6.0
kombu>=5.4,<6.0
redis>=5.0,<6.0
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
requests>=2.31,<3.0

